[release test] make more files a standalone py_library's
working towards breaking down `ray_release package` into
smaller modules

moves ExitCode to exception.py
