+++ release/ray_release/exception.py
@@ -1,4 +1,38 @@
-from ray_release.result import ExitCode
+import enum
+
+
+class ExitCode(enum.Enum):
+    # If you change these, also change the `retry` section
+    # in `build_pipeline.py` and the `reason()` function in `run_e2e.sh`
+    SUCCESS = 0  # Do not set/return this manually
+    UNCAUGHT = 1  # Do not set/return this manually
+
+    UNSPECIFIED = 2
+    UNKNOWN = 3
+
+    # Hard infra errors (non-retryable)
+    CLI_ERROR = 10
+    CONFIG_ERROR = 11
+    SETUP_ERROR = 12
+    CLUSTER_RESOURCE_ERROR = 13
+    CLUSTER_ENV_BUILD_ERROR = 14
+    CLUSTER_STARTUP_ERROR = 15
+    LOCAL_ENV_SETUP_ERROR = 16
+    REMOTE_ENV_SETUP_ERROR = 17
+    FETCH_RESULT_ERROR = 18
+    ANYSCALE_ERROR = 19
+
+    # Infra timeouts (retryable)
+    RAY_WHEELS_TIMEOUT = 30
+    CLUSTER_ENV_BUILD_TIMEOUT = 31
+    CLUSTER_STARTUP_TIMEOUT = 32
+    CLUSTER_WAIT_TIMEOUT = 33
+
+    # Command errors - these are considered application errors
+    COMMAND_ERROR = 40
+    COMMAND_ALERT = 41
+    COMMAND_TIMEOUT = 42
+    PREPARE_ERROR = 43
 
 
 class ReleaseTestError(RuntimeError):
