+++ release/ray_release/result.py
@@ -3,6 +3,8 @@ import os
 from dataclasses import dataclass
 from typing import Optional, Dict, Tuple
 
+from ray_release.exception import ExitCode, ReleaseTestError
+
 
 class ResultStatus(enum.Enum):
     """
@@ -45,41 +47,7 @@ class Result:
     extra_tags: Optional[Dict] = None
 
 
-class ExitCode(enum.Enum):
-    # If you change these, also change the `retry` section
-    # in `build_pipeline.py` and the `reason()` function in `run_e2e.sh`
-    SUCCESS = 0  # Do not set/return this manually
-    UNCAUGHT = 1  # Do not set/return this manually
-
-    UNSPECIFIED = 2
-    UNKNOWN = 3
-
-    # Hard infra errors (non-retryable)
-    CLI_ERROR = 10
-    CONFIG_ERROR = 11
-    SETUP_ERROR = 12
-    CLUSTER_RESOURCE_ERROR = 13
-    CLUSTER_ENV_BUILD_ERROR = 14
-    CLUSTER_STARTUP_ERROR = 15
-    LOCAL_ENV_SETUP_ERROR = 16
-    REMOTE_ENV_SETUP_ERROR = 17
-    FETCH_RESULT_ERROR = 18
-    ANYSCALE_ERROR = 19
-
-    # Infra timeouts (retryable)
-    RAY_WHEELS_TIMEOUT = 30
-    CLUSTER_ENV_BUILD_TIMEOUT = 31
-    CLUSTER_STARTUP_TIMEOUT = 32
-    CLUSTER_WAIT_TIMEOUT = 33
-
-    # Command errors - these are considered application errors
-    COMMAND_ERROR = 40
-    COMMAND_ALERT = 41
-    COMMAND_TIMEOUT = 42
-    PREPARE_ERROR = 43
-
-
-def _is_transient_error(result_status: ResultStatus, runtime: int) -> bool:
+def _is_transient_error(runtime: int) -> bool:
     """
     Classify whether an infra-failure issue is a transient issue. This is based on
     the status of its previous retries, and its runtime.
@@ -98,7 +66,6 @@ def _is_transient_error(result_status: ResultStatus, runtime: int) -> bool:
 def handle_exception(
     e: Exception, run_duration: int
 ) -> Tuple[ExitCode, ResultStatus, Optional[int]]:
-    from ray_release.exception import ReleaseTestError
 
     if not isinstance(e, ReleaseTestError):
         return ExitCode.UNKNOWN, ResultStatus.UNKNOWN, 0
@@ -121,7 +88,7 @@ def handle_exception(
 
     # if this result is to be retried, mark its status as transient
     # this logic should be in-sync with run_release_test.sh
-    if _is_transient_error(result_status, run_duration):
+    if _is_transient_error(run_duration):
         result_status = ResultStatus.TRANSIENT_INFRA_ERROR
 
     return exit_code, result_status, runtime
