+++ release/ray_release/tests/test_glue.py
@@ -18,6 +18,7 @@ from ray_release.exception import (
     ClusterCreationError,
     ClusterStartupError,
     ClusterStartupTimeout,
+    ExitCode,
     RemoteEnvSetupError,
     CommandError,
     PrepareCommandError,
@@ -38,7 +39,7 @@ from ray_release.glue import (
 )
 from ray_release.logger import logger
 from ray_release.reporter.reporter import Reporter
-from ray_release.result import Result, ExitCode
+from ray_release.result import Result
 from ray_release.tests.utils import MockSDK, APIDict
