+++ release/ray_release/tests/test_result.py
@@ -1,9 +1,11 @@
-import pytest
 import sys
 import os
 from unittest import mock
-from ray_release.result import handle_exception, ExitCode, ResultStatus
-from ray_release.exception import ReleaseTestError, ReleaseTestSetupError
+
+import pytest
+
+from ray_release.result import handle_exception, ResultStatus
+from ray_release.exception import ExitCode, ReleaseTestError, ReleaseTestSetupError
 
 
 def test_handle_exception():
